test: reuse authoritative SSH connection readiness in localhost fixture
